Added switching between audio and video
In particular, I updated what I programmed yesterday and copied our prior code (for playing audio) into AugmentedImageActivity. Arguably, the most important feature that was changed is the location of where audio is created and started, which is now under the hashmap check (to see if an image has been detected before). I created a new list of booleans (in AugmentedImageFragment, similar to videoList and musicList) to check if a video should be played or music should be played. The program checks if a video should be played first (AugmentedImageFragment.imagePlaysVideoBooleanList[augmentedImage.getIndex() is true), otherwise, an audio file is played.